requirements.txt
Django>=1.10,<1.11
channels
asgi_redis
redis
psycopg2
Pillow
gunicorn
requests
beautifulsoup4
simplejson
django-haystack
elasticsearch>=2.0.0,<3.0.0
djangorestframework
djangorestframework-jwt
django-cors-headers
django-crispy-forms
django-imagekit
django-annoying
django-simplemde
python-slugify
django-tables2
django-columns
django-recaptcha
django-extensions
django-ajax-selects
django-multiselectfield
#python-memcached
mongoengine
hashids
cssselect
lxml
unidecode
pytz
arrow
whenpy
python-Levenshtein
fuzzywuzzy
Celery
flower
click
ptpython
markdown
markdown2
pycountry
xmljson
ftfy
html2text
tldextract
mammoth
radar
httpagentparser